Move systray code into separate CU 	modified:   src/CMakeLists.txt 	modified:   src/mainwindow.cc 	modified:   src/mainwindow.h 	new file:   src/systray.cc 	new file:   src/systray.h
